feat: add allProjects action to retrieve all projects from the database
This commit adds a new action called allProjects to the revver123/actions directory. The action uses the db.select() method to retrieve all projects from the projects table in the database. The retrieved projects are then returned by the action. This change enables the retrieval of all projects for further processing or display.