feat: Add createProject action to create a new project
The createProject action is added to the revver123/actions directory. This action allows users to create a new project by providing the title, URL, and description. The action retrieves the authenticated user's ID and creates a new project object with the provided data. The project is then inserted into the database using the db.insert() method. The newly created project is returned by the action.