feat: Add slug column to projects table
The code changes add a new column called "slug" to the "projects" table in the database schema. This column will store a URL-friendly version of the project title. The slugify() function is also added to the utils.js file to generate the slug based on the project title. This change enables the creation of SEO-friendly URLs for projects.