feat: Add CopyBtn component for copying code in integrated it in project instructions page.
The code changes add a new component called CopyBtn.jsx, which is responsible for rendering a button that allows users to copy code to the clipboard. The component uses the navigator.clipboard API to write the code to the clipboard when the button is clicked. It also includes a tooltip that provides a brief description of the button's functionality. This change enhances the user experience by providing an easy way to copy code snippets.